+++ src/union_physics/v8/edgebreaker.rs
@@ -91,7 +91,7 @@ impl<'a> HullDecoder<'a> {
 			vertex_offset: 0,
 		}
 	}
-	pub fn remaining_bits(&self) -> usize {
+	pub fn remaining_bits(&self) -> u32 {
 		self.symbol_reader.remaining_bits()
 	}
 	pub fn current_face(&self) -> u32 {
